+++ src/mousedroid/telemetry/auth.py
@@ -45,7 +45,7 @@ def build_bearer_auth_middleware(
 
     token = os.environ.get(auth_cfg.token_env_var, "")
 
-    @web.middleware  # type: ignore[misc,untyped-decorator,unused-ignore]
+    @web.middleware
     async def bearer_auth_middleware(
         request: web.Request,
         handler: Any,
@@ -128,7 +128,7 @@ def build_cors_middleware(
     """
     from aiohttp import web
 
-    @web.middleware  # type: ignore[misc,untyped-decorator,unused-ignore]
+    @web.middleware
     async def cors_middleware(
         request: web.Request,
         handler: Any,